+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 
 ### Changes:
 
+-   :boom: Breaking Changes
+
+    -   The default label alignment for bots in the page portal has changed from `center` to `left`. It is now consistent across all portals.
+
 -   :rocket: Improvements
 
     -   Added the `player.share(options)` function.
@@ -18,8 +22,8 @@
     -   Added the `auxLabelAlignment` tag.
         -   Note that this value affects menu bots as well.
         -   Possible values are:
-            -   `center` - Aligns the text in the center of the label. (default)
-            -   `left` - Aligns the text to the left of the label.
+            -   `left` - Aligns the text to the left of the label. (default)
+            -   `center` - Aligns the text in the center of the label.
             -   `right` - Aligns the text to the right of the label.
 
 ## V1.0.26